+++ pkg/deployment/reconcile/plan_builder_test.go
@@ -508,11 +508,13 @@ func TestCreatePlanActiveFailoverScale(t *testing.T) {
 	}
 	newPlan, changed = createNormalPlan(ctx, log, depl, nil, spec, status, inspector.NewEmptyInspector(), c)
 	assert.True(t, changed)
-	require.Len(t, newPlan, 2) // Note: Downscaling is only down 1 at a time
-	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[0].Type)
-	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[1].Type)
+	require.Len(t, newPlan, 3) // Note: Downscaling is only down 1 at a time
+	assert.Equal(t, api.ActionTypeKillMemberPod, newPlan[0].Type)
+	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[1].Type)
+	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[2].Type)
 	assert.Equal(t, api.ServerGroupSingle, newPlan[0].Group)
 	assert.Equal(t, api.ServerGroupSingle, newPlan[1].Group)
+	assert.Equal(t, api.ServerGroupSingle, newPlan[2].Group)
 }
 
 // TestCreatePlanClusterScale creates a `cluster` deployment to test the creating of scaling plan.
@@ -610,17 +612,21 @@ func TestCreatePlanClusterScale(t *testing.T) {
 	spec.Coordinators.Count = util.NewInt(1)
 	newPlan, changed = createNormalPlan(ctx, log, depl, nil, spec, status, inspector.NewEmptyInspector(), c)
 	assert.True(t, changed)
-	require.Len(t, newPlan, 5) // Note: Downscaling is done 1 at a time
+	require.Len(t, newPlan, 7) // Note: Downscaling is done 1 at a time
 	assert.Equal(t, api.ActionTypeCleanOutMember, newPlan[0].Type)
-	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[1].Type)
-	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[2].Type)
-	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[3].Type)
-	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[4].Type)
+	assert.Equal(t, api.ActionTypeKillMemberPod, newPlan[1].Type)
+	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[2].Type)
+	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[3].Type)
+	assert.Equal(t, api.ActionTypeKillMemberPod, newPlan[4].Type)
+	assert.Equal(t, api.ActionTypeShutdownMember, newPlan[5].Type)
+	assert.Equal(t, api.ActionTypeRemoveMember, newPlan[6].Type)
 	assert.Equal(t, api.ServerGroupDBServers, newPlan[0].Group)
 	assert.Equal(t, api.ServerGroupDBServers, newPlan[1].Group)
 	assert.Equal(t, api.ServerGroupDBServers, newPlan[2].Group)
-	assert.Equal(t, api.ServerGroupCoordinators, newPlan[3].Group)
+	assert.Equal(t, api.ServerGroupDBServers, newPlan[3].Group)
 	assert.Equal(t, api.ServerGroupCoordinators, newPlan[4].Group)
+	assert.Equal(t, api.ServerGroupCoordinators, newPlan[5].Group)
+	assert.Equal(t, api.ServerGroupCoordinators, newPlan[6].Group)
 }
 
 type LastLogRecord struct {
@@ -793,6 +799,7 @@ func TestCreatePlan(t *testing.T) {
 				ad.Status.Members.Agents[0].PersistentVolumeClaimName = pvcName
 			},
 			ExpectedPlan: []api.Action{
+				api.NewAction(api.ActionTypeKillMemberPod, api.ServerGroupAgents, ""),
 				api.NewAction(api.ActionTypeShutdownMember, api.ServerGroupAgents, ""),
 				api.NewAction(api.ActionTypeRemoveMember, api.ServerGroupAgents, ""),
 				api.NewAction(api.ActionTypeAddMember, api.ServerGroupAgents, ""),
@@ -989,6 +996,7 @@ func TestCreatePlan(t *testing.T) {
 			},
 			ExpectedPlan: []api.Action{
 				api.NewAction(api.ActionTypeCleanOutMember, api.ServerGroupDBServers, "id"),
+				api.NewAction(api.ActionTypeKillMemberPod, api.ServerGroupDBServers, ""),
 				api.NewAction(api.ActionTypeShutdownMember, api.ServerGroupDBServers, ""),
 				api.NewAction(api.ActionTypeRemoveMember, api.ServerGroupDBServers, ""),
 			},
@@ -1007,6 +1015,7 @@ func TestCreatePlan(t *testing.T) {
 			},
 			ExpectedPlan: []api.Action{
 				api.NewAction(api.ActionTypeCleanOutMember, api.ServerGroupDBServers, "id"),
+				api.NewAction(api.ActionTypeKillMemberPod, api.ServerGroupDBServers, ""),
 				api.NewAction(api.ActionTypeShutdownMember, api.ServerGroupDBServers, ""),
 				api.NewAction(api.ActionTypeRemoveMember, api.ServerGroupDBServers, ""),
 			},