+++ pkg/deployment/resources/pod_finalizers.go
@@ -78,6 +78,12 @@ func (r *Resources) runPodFinalizers(ctx context.Context, p *v1.Pod, memberStatu
 			} else {
 				log.Debug().Err(err).Str("finalizer", f).Msg("Cannot remove Pod finalizer yet")
 			}
+		case constants.FinalizerPodGracefulShutdown:
+			// We are in graceful shutdown, only one way to remove it is when container is already dead
+			if isServerContainerDead {
+				log.Debug().Msg("Server Container is dead, removing finalizer")
+				removalList = append(removalList, f)
+			}
 		case constants.FinalizerDelayPodTermination:
 			if isServerContainerDead {
 				log.Debug().Msg("Server Container is dead, removing finalizer")
@@ -93,11 +99,12 @@ func (r *Resources) runPodFinalizers(ctx context.Context, p *v1.Pod, memberStatu
 			log.Error().Str("finalizer", f).Msg("Delay finalizer")
 
 			groupSpec := r.context.GetSpec().GetServerGroupSpec(group)
-			d := time.Duration(util.IntOrDefault(groupSpec.ShutdownDelay, 0)) * time.Second
 			if t := p.ObjectMeta.DeletionTimestamp; t != nil {
-				e := p.ObjectMeta.DeletionTimestamp.Time.Sub(time.Now().Add(d))
-				log.Error().Str("finalizer", f).Dur("left", e).Msg("Delay finalizer status")
-				if e < 0 {
+				d := time.Duration(groupSpec.GetShutdownDelay(group)) * time.Second
+				gr := time.Duration(util.Int64OrDefault(p.ObjectMeta.GetDeletionGracePeriodSeconds(), 0)) * time.Second
+				e := t.Time.Add(-1 * gr).Sub(time.Now().Add(-1 * d))
+				log.Error().Str("finalizer", f).Str("left", e.String()).Msg("Delay finalizer status")
+				if e < 0 || d == 0 {
 					removalList = append(removalList, f)
 				}
 			} else {